+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 - Added support for manage a single serial device
 - Added support for manage multiple serial devices
 - Added a socketio client
+
+**Release 0.1.0.a** - 2022-06-18
+- Added an MJPEG server
+- Added JPEG encoder on Camera API